% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difLR.R
\name{difLR}
\alias{difLR}
\title{Differential Item Functioning Study}
\usage{
difLR(x, group, focal, complete = TRUE, na.rm = FALSE)
}
\arguments{
\item{x}{matrix or data.frame of scored item responses.}

\item{complete}{logical with default \code{TRUE} indicating whether or not
\code{x} should be reduced to rows with complete data across all columns.}

\item{na.rm}{logical with default \code{FALSE} specifying whether missings
should be removed before calculating individual descriptives.}

\item{subset}{optional vector for selecting a subset of columns from \code{x}.}

\item{scores}{optional vector of construct scores.}

\item{groups}{optional grouping variable.}
}
\description{
Functions for estimating differential item functioning with a set of
scored item responses.
}

